# renren-fast/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: renren-fast是一個轻量级的Java快速開發平台，能快速開發项目并交付【接私活利器】
  version: 1.0.0
  title: 人人快速開發平台
basePath: /renren-fast
schemes:
  - http

#認證
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定義介面資料
paths:
...
  /sys/user/list:
    get:
      tags:
        - 管理員管理
      summary: 管理員列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序欄位
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: username
          description: 使用者名
          in: query
          type: string
      responses:
        '200':
          description: 返回管理員列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
...
  /sys/user/info/{userId}:
    get:
      tags:
        - 管理員管理
      summary: 取得管理員信息
      produces:
        - application/json
      parameters:
        - name: userId
          description: 使用者ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回管理員信息
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserEntity'
  /sys/user/password:
    post:
      tags:
        - 管理員管理
      summary: 修改密碼
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理員物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/PasswordForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/save:
    post:
      tags:
        - 管理員管理
      summary: 新增管理員
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理員物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/update:
    post:
      tags:
        - 管理員管理
      summary: 修改管理員
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理員物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/delete:
    post:
      tags:
        - 管理員管理
      summary: 刪除管理員
      produces:
        - application/json
      parameters:
        - name: body
          description: 使用者ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/list:
    get:
      tags:
        - 角色管理
      summary: 角色列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序欄位
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: roleName
          description: 角色名
          in: query
          type: string
      responses:
        '200':
          description: 返回角色列表
          schema:
            $ref: '#/definitions/SysRoleEntityList'
...
  /sys/role/info/{roleId}:
    get:
      tags:
        - 角色管理
      summary: 取得角色信息
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回角色信息
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              role:
                $ref: '#/definitions/SysRoleEntity'
  /sys/role/save:
    post:
      tags:
        - 角色管理
      summary: 新增角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/update:
    post:
      tags:
        - 角色管理
      summary: 修改角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/delete:
    post:
      tags:
        - 角色管理
      summary: 刪除角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/menu/info/{menuId}:
    get:
      tags:
        - 選單管理
      summary: 取得選單信息
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 選單ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回選單信息
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              menu:
                description: 選單信息
                $ref: '#/definitions/SysMenuEntity'
  /sys/menu/save:
    post:
      tags:
        - 選單管理
      summary: 新增選單
      produces:
        - application/json
      parameters:
        - name: body
          description: 選單物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
        - 選單管理
      summary: 修改選單
      produces:
        - application/json
      parameters:
        - name: body
          description: 選單物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/delete/{menuId}:
    post:
      tags:
        - 選單管理
      summary: 刪除選單
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 選單ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 狀態碼 0：成功  非0：失敗
          type: integer
          format: int32
        msg:
          description: 失敗原因
          type: string
...
  PasswordForm:
    type: object
    properties:
      password:
        description: 原密碼
        type: string
      newPassword:
        description: 新密碼
        type: string
  SysUserEntity:
    type: object
    properties:
      userId:
        description: 使用者ID
        type: integer
        format: int64
      username:
        description: 使用者名
        type: string
      password:
        description: 密碼
        type: string
      email:
        description: E-mail信箱
        type: string
      mobile:
        description: 手機
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 建立者ID
        type: integer
        format: int64
      createTime:
        description: 建立时间
        type: string
        format: date-time
  SysUserEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录數
            type: integer
            format: int32
          pageSize:
            description: 每页记录數
            type: integer
            format: int32
          totalPage:
            description: 总页數
            type: integer
            format: int32
          currPage:
            description: 目前页數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserEntity'
  SysUserEntityEdit:
      type: object
      properties:
        userId:
          description: 使用者ID
          type: integer
          format: int64
        username:
          description: 使用者名
          type: string
        password:
          description: 密碼
          type: string
        email:
          description: E-mail信箱
          type: string
        mobile:
          description: 手機
          type: string
        status:
          description: 状态  0：禁用   1：正常
          type: integer
          format: int32
        roleIdList:
          description: 角色ID列表
          type: array
          items:
            type: integer
            format: int32
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名稱
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 選單ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 建立者ID
        type: integer
        format: int64
      createTime:
        description: 建立时间
        type: string
        format: date-time
  SysRoleEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录數
            type: integer
            format: int32
          pageSize:
            description: 每页记录數
            type: integer
            format: int32
          totalPage:
            description: 总页數
            type: integer
            format: int32
          currPage:
            description: 目前页數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysRoleEntity'
  SysRoleEntityEdit:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名稱
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 選單ID列表
        type: array
        items:
          type: integer
          format: int64
  SysMenuEntity:
    type: object
    properties:
      menuId:
        description: 選單ID
        type: integer
        format: int64
      name:
        description: 選單名稱
        type: string
      parentId:
        description: 父選單ID，一級選單為0
        type: integer
        format: int64
      parentName:
        description: 父選單名稱
        type: string
      url:
        description: 選單URL
        type: string
      perms:
        description: 授權标识
        type: string
      type:
        description: 類型     0：目录   1：選單   2：按鈕
        type: integer
        format: int32
      icon:
        description: 選單图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展开   true：展开   false：不展开
        type: boolean
        format: int32
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 選單ID
        type: integer
        format: int64
      name:
        description: 選單名稱
        type: string
      parentId:
        description: 父選單ID，一級選單為0
        type: integer
        format: int64
      url:
        description: 選單URL
        type: string
      perms:
        description: 授權标识
        type: string
      type:
        description: 類型     0：目录   1：選單   2：按鈕
        type: integer
        format: int32
      icon:
        description: 選單图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
...